Ignore out of source build directory for cmake.
The common pattern with cmake from the top level directory is:

    mkdir build
    cd build
    cmake ../

This is typically referred to as an out of source build. This leaves
your project tree intact and does not pollute it with build output
files. As a result, it makes sense to exclude this directory so that
it doesn't accidentally get checked in to git.